refactor: reorganize auth JWT modules and centralize frontend types
Align backend auth imports with new jwt submodule structure and move shared frontend interfaces into a common types layer while updating store and middleware imports.
