+++ apps/backend/src/modules/auth/dto/response/login.response.ts
@@ -1,0 +1,10 @@
+export class LoginResponse {
+    user: {
+        id: string,
+        email: string,
+        displayName: string | null,
+        photoUrl: string | null
+    }
+    token: string
+    refreshToken: string
+}
