+++ apps/backend/src/modules/auth/jwt/jwt.strategy.ts
