+++ apps/backend/src/modules/conversations/conversations.controller.ts
@@ -145,7 +145,7 @@ export class ConversationsController {
         statusCode: HttpStatus.BAD_REQUEST,
       })
     }
-
+// set SSE
     res.setHeader('Content-Type', 'text/event-stream')
     res.setHeader('Cache-Control', 'no-cache')
     res.setHeader('Connection', 'keep-alive')
