+++ apps/frontend/src/auth/RequireAuth.tsx
@@ -1,14 +1,15 @@
-import { Navigate, Outlet } from 'react-router-dom'
+import { Navigate, Outlet, useLocation } from 'react-router-dom'
 import { useAuth } from '../hooks/useAuth'
 import { useEffect } from 'react'
 
 const RequireAuth = () => {
-  const { isAuthenticated, initialize } = useAuth()
+  const { isAuthenticated, initialize } = useAuth();
+  const location = useLocation();
 
   // Initialize auth state on component mount
   useEffect(() => {
     initialize()
-  }, [initialize])
+  }, [location])
 
   if (!isAuthenticated) {
     return <Navigate to="/login" replace state={{ from: window.location.pathname }} />
