+++ apps/frontend/src/store/store.ts
@@ -1,9 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit'
-import conversationReducer from './slices/conversation.slice'
-import messageReducer from './slices/message.slice'
-import authReducer from './slices/auth.slice'
-import userReducer from './slices/user.slice'
-import themeReducer from './slices/theme.slice'
+import conversationReducer from './slices/conversation.slice.ts'
+import messageReducer from './slices/message.slice.ts'
+import authReducer from './slices/auth.firebase.slice.ts'
+import userReducer from './slices/user.slice.ts'
+import themeReducer from './slices/theme.slice.ts'
 
 export const store = configureStore({
   reducer: {
